# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Catalog of indie games — browser or download, free, paid or subscription.",
        "title": "IndieForge API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.AuthResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign out",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.UserDTO"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new account",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.AuthResponse"
                        }
                    }
                }
            }
        },
...
        "/games/{id}/claim-free": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commerce"
                ],
                "summary": "Add a free (or demo-active) game to the library",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game id or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/indieforge_internal_dto.GameDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/games/{id}/perks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commerce"
                ],
                "summary": "Subscriber chat link (subscribers/author only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game id or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/me/library": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commerce"
                ],
                "summary": "Owned + subscribed games",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/payments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commerce"
                ],
                "summary": "Start a YooKassa payment (purchase | friend-pack | subscription)",
                "parameters": [
                    {
                        "description": "Payment request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/indieforge_internal_dto.PaymentDTO"
                        }
                    }
                }
            }
        },
        "/payments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commerce"
                ],
                "summary": "Payment status (poll after redirect)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/payments/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "commerce"
                ],
                "summary": "Cancel a pending payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "indieforge_internal_dto.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/indieforge_internal_dto.UserDTO"
                }
            }
        },
        "indieforge_internal_dto.CreatePaymentRequest": {
            "type": "object",
            "properties": {
                "friendUsername": {
                    "type": "string"
                },
                "gameId": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
...
        "indieforge_internal_dto.DemoDayDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "endsAt": {
                    "type": "string"
                },
                "startsAt": {
                    "type": "string"
                }
            }
        },
        "indieforge_internal_dto.GameDTO": {
            "type": "object",
            "properties": {
                "browserBuildUrl": {
                    "type": "string"
                },
                "canPlayFree": {
                    "type": "boolean"
                },
                "coverImage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "demoDay": {
                    "$ref": "#/definitions/indieforge_internal_dto.DemoDayDTO"
                },
                "description": {
                    "type": "string"
                },
                "developerId": {
                    "type": "string"
                },
                "developerName": {
                    "type": "string"
                },
                "downloadFileName": {
                    "type": "string"
                },
                "downloadPlatforms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "downloadSizeMB": {
                    "type": "integer"
                },
                "friendPackDiscount": {
                    "type": "integer"
                },
                "genre": {
                    "type": "string"
                },
                "hasBrowserBuild": {
                    "type": "boolean"
                },
                "hasDownloadBuild": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "owned": {
                    "type": "boolean"
                },
                "price": {
                    "type": "integer"
                },
                "pricingModel": {
                    "type": "string"
                },
                "screenshots": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "stats": {
                    "$ref": "#/definitions/indieforge_internal_dto.StatsDTO"
                },
                "status": {
                    "type": "string"
                },
                "subscribed": {
                    "type": "boolean"
                },
                "subscription": {
                    "$ref": "#/definitions/indieforge_internal_dto.SubscriptionDTO"
                },
                "supportsMultiplayer": {
                    "type": "boolean"
                },
                "tagline": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "theme": {
                    "$ref": "#/definitions/indieforge_internal_dto.Theme"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "indieforge_internal_dto.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "indieforge_internal_dto.PaymentDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "confirmationUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "friendUsername": {
                    "type": "string"
                },
                "gameId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "indieforge_internal_dto.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "indieforge_internal_dto.StatsDTO": {
            "type": "object",
            "properties": {
                "owners": {
                    "type": "integer"
                },
                "plays": {
                    "type": "integer"
                },
                "subscribers": {
                    "type": "integer"
                }
            }
        },
        "indieforge_internal_dto.SubscriptionDTO": {
            "type": "object",
            "properties": {
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "enabled": {
                    "type": "boolean"
                },
                "period": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "indieforge_internal_dto.Theme": {
            "type": "object",
            "properties": {
                "accent": {
                    "type": "string"
                },
                "accent2": {
                    "type": "string"
                },
                "background": {
                    "type": "string"
                },
                "cardShape": {
                    "type": "string"
                },
                "layout": {
                    "type": "string"
                }
            }
        },
        "indieforge_internal_dto.UserDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDeveloper": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}